HID: report_timeout and sony Sixaxis shanwan clone controllers have an issue with rumble. It works, but every rumble event continues for 10 seconds and then all rumble events do not work for  the follwing 10 seconds after that.
After extensive testing and debugging, I was able to identify that sending a soft-reset
to the controller will stop the 10 second rumble, and then that the delays are due
to a timeout waiting for HIDP_WAITING_FOR_SEND_ACK.

So I've added a report-timeout member to the hid device struct. I can then use this
in the sony hid driver to make rumble work properly.

Also changed the shanwan quirk to be specific to usb and added a separate bluetooth
quirk for the new report timeout fix.

Changed the shanwan usb quirk to check for any usb sony gamepad with shanwan in the
name. This works well with a new bluez patch that pulls the udev pad name when paired.

For completeness, I've added the DualShock3 rumble duration to the report data. It
wasn't set before, and the ff memless uses the same value but let's let the gamepad
know the value too.

Lastly, add a quirk to support Gasia devices that are all named the same.