input: xpad: updates, init sequences, and shanwan Include some updates from upstream, as well as a couple tweaks from paraj's fork. This includes adding code to activate some clone controllers. Without this activation sequence, no inputs are recognized.
This activation sequence breaks some controllers, however. So I have changed the mapping property in the driver to be quirks instead and added a quirk to enable this activation sequence. For now I have only enabled the sequence for a couple controllers - because I have them and can test them.

Unfortunately - the way some of these clones work is that they switch firmware after being initialized. When they are re-initialized they are detected as normal XBOX360 controllers with completely different productid and vendorid. So the quirks defined are then lost.

As a workaround, I've added a manufacturer check so that even when a clone uses the standard xbox360 product/vendor, the manufacturer is inspected and these quirks can be enabled.

I've applied the needed quirks for the SHANWAN manufacturer during probe, and I also discoevered these shanwan clones need USB reset after resume. I wonder if there's any harm in just setting that for all devices?