rockchip: dmc: stop thermals from disabling rockchip dmc freqs
Thermal throttling the memory on an SBC can really affect performance,
especially during 4K HD playback.

The rockchip dmc driver uses system-status-freq to auto-select a frequency
based on system state, such as playing 4K / 1080p videos or during suspend.
This is already a type of throttle.

Thermals will still throttle the CPU and usually GPU, so let's leave the
memory alone. One option would be to adjust the thermal degree limits for
the chipset in the device tree, but that would involve a lot of trial
and error. So let's just disable it.