4.4.302 This is officially EOL for the 4.4 kernel.
RIP :`(

I will continue my current process of backporting changes from
upsream in order to support select rockchip devices. But I will also
begin the process of forking mainline and working towards proper
support for these devices in a new repo. Sad to say this is the
beginning of the end for 4.4 BSP